correct arboresence of files and create a folder for each pages
