+++ src/pages/Connexion/Connexion.tsx
@@ -1,0 +1,62 @@
+import Button from "@mui/material/Button";
+import TextField from "@mui/material/TextField";
+import { useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { loginUser } from "../../services/api/auth";
+import Header from "../../components/Header";
+function Connexion() {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const navigate = useNavigate();
+
+  const handleSubmit = async (event: React.FormEvent) => {
+    event.preventDefault();
+    try {
+      const data = await loginUser(email, password);
+      localStorage.setItem("auth_token", data.token);
+      navigate("/dashboard");
+    } catch (err) {
+      console.error("Erreur lors de la connexion :", err);
+    }
+  };
+  return (
+    <>
+      <Header />
+      <div className="flex min-h-full flex-1 flex-col justify-center px-6 py-12 lg:px-8">
+        <div className="sm:mx-auto sm:w-full sm:max-w-sm">
+          <h2 className="mt-10 text-center text-2xl/9 font-bold tracking-tight text-gray-900">
+            Sign in to your account
+          </h2>
+        </div>
+
+        <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-sm">
+          <form onSubmit={handleSubmit} method="POST" className="space-y-6">
+            <TextField
+              id="outlined-basic"
+              label="Email adress"
+              variant="outlined"
+              type="email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              required
+            />
+            <TextField
+              id="outlined-basic"
+              label="Password"
+              variant="outlined"
+              type="password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              required
+            />
+            <Button variant="outlined" type="submit">
+              Connexion
+            </Button>
+          </form>
+        </div>
+      </div>
+    </>
+  );
+}
+
+export default Connexion;
