+++ src/pages/Home/Home.tsx
@@ -1,0 +1,12 @@
+import DrawerAppBar from "../../components/Header";
+import Slider from "../../components/Slider";
+function Home() {
+  return (
+    <>
+      <DrawerAppBar />
+      <Slider />
+    </>
+  );
+}
+
+export default Home;
